# frontend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ShiguReader API",
    "description": "ShiguReader backend API — 类似 FastAPI 的交互式文档",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
      "FileSystemItem": {
        "title": "FileSystemItem",
        "type": "object",
        "required": [
          "name",
          "path",
          "item_type"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "item_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "folder"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "file"
                ]
              }
            ]
          },
          "file_type": {
            "anyOf": [
              {
                "anyOf": [
                  {
                    "type": "string",
                    "enum": [
                      "image"
                    ]
                  },
                  {
                    "type": "string",
                    "enum": [
                      "video"
                    ]
                  },
                  {
                    "type": "string",
                    "enum": [
                      "archive"
                    ]
                  },
                  {
                    "type": "string",
                    "enum": [
                      "audio"
                    ]
                  },
                  {
                    "type": "string",
                    "enum": [
                      "unknown"
                    ]
                  }
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "filesize": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "mtime": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "likeScore": {
            "default": 0,
            "type": "number"
          },
          "is_missing": {
            "default": 0,
            "type": "integer"
          },
          "image_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "video_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "audio_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "avg_image_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_read_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
      "LocalCheckHit": {
        "title": "LocalCheckHit",
        "type": "object",
        "required": [
          "name",
          "path",
          "thumbnail_url",
          "match_level",
          "confidence"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "match_level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "downloaded"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "likely"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "same_author"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "different"
                ]
              }
            ]
          },
          "confidence": {
            "type": "number"
          }
        }
      },
      "LocalCheckResult": {
        "title": "LocalCheckResult",
        "type": "object",
        "required": [
          "q",
          "match_level",
          "confidence",
          "reason",
          "hits"
        ],
        "properties": {
          "q": {
            "type": "string"
          },
          "match_level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "downloaded"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "likely"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "same_author"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "different"
                ]
              }
            ]
          },
          "confidence": {
            "type": "number"
          },
          "reason": {
            "type": "string"
          },
          "hits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LocalCheckHit"
            }
          }
        }
      },
...
      "ParseResponse": {
        "title": "ParseResponse",
        "type": "object",
        "required": [
          "title",
          "authors",
          "raw_tags",
          "type"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "authors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "cosers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "raw_tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "event": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "date_tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "type": {
            "type": "string"
          },
          "pack_kind": {
            "default": "manga",
            "type": "string"
          }
        }
      },
...
      "BatchParseItem": {
        "title": "BatchParseItem",
        "type": "object",
        "required": [
          "filepath"
        ],
        "properties": {
          "filepath": {
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParseResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/search": {
      "post": {
        "operationId": "searchFiles",
        "summary": "搜索文件（支持文件名/作者/coser/标签）",
        "tags": [
          "Search"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "title": "SearchRequest",
                "type": "object",
                "properties": {
                  "q": {
                    "default": "",
                    "type": "string"
                  },
                  "scopes": {
                    "type": "array",
                    "items": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "file"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "author"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "coser"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "tag"
                          ]
                        }
                      ]
                    }
                  },
                  "mode": {
                    "default": "fuzzy",
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "exact"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "fuzzy"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "local-check"
                        ]
                      }
                    ]
                  },
                  "presence_filter": {
                    "default": "all",
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "all"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "watched"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "scanned_recent"
                        ]
                      }
                    ]
                  },
                  "limit": {
                    "default": 200,
                    "type": "integer"
                  },
                  "offset": {
                    "default": 0,
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "SearchResponse",
                  "type": "object",
                  "required": [
                    "items",
                    "total"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/FileSystemItem"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/local-check-batch": {
      "post": {
        "operationId": "localCheckBatch",
        "summary": "批量本地持有检查（油猴脚本用）",
        "tags": [
          "Search"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "title": "LocalCheckBatchRequest",
                "type": "object",
                "properties": {
                  "queries": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "limit": {
                    "default": 5,
                    "type": "integer"
                  },
                  "presence_filter": {
                    "default": "all",
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "all"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "watched"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "scanned_recent"
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "LocalCheckBatchResponse",
                  "type": "object",
                  "required": [
                    "results"
                  ],
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/LocalCheckResult"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/parse": {
      "get": {
        "operationId": "getParseResult",
        "summary": "解析单个文件名元数据（作者/标签/标题）",
        "tags": [
          "Parse"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "filepath",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "StoredParseResponse",
                  "type": "object",
                  "required": [
                    "filepath"
                  ],
                  "properties": {
                    "filepath": {
                      "type": "string"
                    },
                    "title": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "authors": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "cosers": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "group_name": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "raw_tags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "event": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "date_tag": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "media_type": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parse/batch": {
      "post": {
        "operationId": "parseBatch",
        "summary": "批量解析文件名元数据",
        "tags": [
          "Parse"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "title": "BatchParseRequest",
                "type": "object",
                "required": [
                  "filepaths"
                ],
                "properties": {
                  "filepaths": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "BatchParseResponse",
                  "type": "object",
                  "required": [
                    "items",
                    "parsed_count",
                    "total_count"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BatchParseItem"
                      }
                    },
                    "parsed_count": {
                      "type": "integer"
                    },
                    "total_count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}